*** Not using subshell to eval the test, and always using diff
When match_method was 'text', I was using a subshell to store the eval
output in a variable, so I could compare with the OK text just using
strings, no files involved. Files were just created if the test failed,
so I could send to diff. As normally the tests pass, it was a great
performance bump in all-inline files.

The main problem using subshells is that context is lost. So simple
tests like:

$ a=1
$ echo $a
1
$

Will fail because the variable $a is not persistent between tests. So
now there's only one way to execute the tests: a plain eval in the
current process, saving the results to a file.

Another problem occur when using the --diff-option option: some diff
options as -i or -w could change the match logic, and my string test
could not follow. So diff is mandatory to perform all the output
comparison. The exception is when using #→ --regex that egrep is used
for the match. But if egrep fails to match, diff is called to show the
results.

We're back in the slow lane, but it's more robust.

Note: $ok_text was renamed to $inline_text.
Note: $match_method is now global.