Support '-' as test file (read from STDIN)
Using `-` to mean that the file comes from from STDIN is a Unix classic.
Let's support it.

The implementation is not that clean, in the sense that the "fake"
temporary path will appear when using two or more input files, and one
of them is `-`. But in the summary table at the end it correctly shows
`-`. Example:

    $ cat x
    $ echo 1
    2
    $

    $ cat x | ./clitest - x
    Testing file /var/folders/zj/1vb37fxj2rd2glwn48nrsjl80000gq/T//clitest.zfwux7/stdin.txt
    #1      echo 1
    -------------------------------------------------
    [FAILED #1, line 1] echo 1
    @@ -1 +1 @@
    -2
    +1
    -------------------------------------------------
    Testing file x
    #2      echo 1
    -------------------------------------------------
    [FAILED #2, line 1] echo 1
    @@ -1 +1 @@
    -2
    +1
    -------------------------------------------------

        ok  fail  skip
        -     1     -    -
        -     1     -    x

    FAIL: 2 of 2 tests failed