Add Makefile, use Docker for lint/test/CI
Instead of installing dependencies in both the local machine and in
Travis runner, install everything inside the clitest Docker image:

- Supported shells: bash, dash, mksh, zsh
- Linter: checkbashisms

The image size grew from ~40MB to ~60MB.

Add a Makefile to isolate all the commands related to dev tasks such as
running linters and tests. Note that by default all those tasks are run
inside the clitest Docker container. To run them directly on the host,
avoiding the container, unset the `docker_run` variable. Examples:

    make test-bash               # test using container's bash
    make test-bash docker_run=   # test using host's bash

The CI configuration (Travis) was updated to avoid installing
dependencies and to run make targets inside a freshly built Docker
container.

Also fixed the `COLUMNS` handling in the tests, so it is compatible with
bash and mksh running inside the container.