Move Docker image to new repo, use dev-only image
In #34, development tools were added to the clitest Docker image, making
it both an end user image and a development environment. This was a bad
idea.

This commit separates them in two different images.

- The public clitest image was moved to a new repository:
  https://github.com/aureliojargas/clitest-docker. Now everything is
  automated and releasing a new clitest image to Docker Hub is just a
  matter of adding a git tag to that repository. Everything related to
  that image was removed from this repository.

- The local development image remains here, named `clitest-dev` and
  described in the new `Dockerfile.dev` file. It does not have clitest
  inside it, it's just the development environment. When running the
  image, the current clitest code should be mounted inside it.

In short, everything dev-related is here. After releasing a new clitest
version (here), go to the other repository to make the official Docker
image for it.