Abort when --post-flight command fails
Previously, any error when running the `--post-flight` command was
ignored. The rationale was that all the "real" tests were already
executed, being the post-flight a mere non-fatal cleanup action.

Now any failure in that step is fatal.
Just like it already is with `--pre-flight`.

The user specifically asked for a cleanup task to be run, he/she should
be notified on any error during its execution. If this is an optional
task for the user, he/she can just add a `|| true` at the end so that it
always succeeds, even in case of errors.