Activate and improve --debug option
The `--debug` option is now functional, since all the calls to
`tt_debug()` were uncommented.

The debug output has changed, now it is hopefully more readable:

- Debug lines are always prefixed by `--`.
- Color changed from dark blue to cyan.
- One identifier has a new name: `LINE_* -> LINE_MISC`
- New a separator line before each new line read from the input test
  file (`INPUT_LINE`), making it easier to spot blocks of messages.

Tests were added for `--debug`.

Note that it's still an undocumented option (it does not appear in
`--help`), because it has no value for the normal user. It's used mostly
to help in finding and fixing bugs in clitest itself.